api/requirements.txt
fastapi
pydantic
uvicorn
numpy
pillow
torch
torchvision
python-multipart
